+++ WalletWasabi.Fluent/Views/Dialogs/CancelTransactionDialogView.axaml
@@ -4,10 +4,10 @@
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              xmlns:c="clr-namespace:WalletWasabi.Fluent.Controls"
              xmlns:converters="clr-namespace:WalletWasabi.Fluent.Converters"
-             xmlns:dialogs="clr-namespace:WalletWasabi.Fluent.ViewModels.Dialogs"
+             xmlns:features="clr-namespace:WalletWasabi.Fluent.ViewModels.Wallets.Home.History.Features"
              mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
              x:Class="WalletWasabi.Fluent.Views.Dialogs.CancelTransactionDialogView"
-             x:DataType="dialogs:CancelTransactionDialogViewModel"
+             x:DataType="features:CancelTransactionDialogViewModel"
              x:CompileBindings="True">
 
   <c:ContentArea Title="Cancel Transaction"
